+++ src/devtools/devtools.ts
@@ -1,0 +1,6 @@
+/**
+ * DevTools page: registers the "Wadjet" panel.
+ *
+ * @module
+ */
+void browser.devtools.panels.create('Wadjet', 'icons/wadjet.svg', 'devtools/panel.html');